refactor background layer selector and rename components
